requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
Flask-SocketIO==5.5.1
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
